+++ src/cpp/session/SessionClientEvent.cpp
@@ -149,6 +149,7 @@ const int kChunkOutput = 125;
 const int kChunkOutputFinished = 126;
 const int kRprofStarted = 127;
 const int kRprofStopped = 128;
+const int kRprofCreated = 129;
 }
 
 void ClientEvent::init(int type, const json::Value& data)
@@ -406,6 +407,8 @@ std::string ClientEvent::typeName() const
          return "rprof_started";
       case client_events::kRprofStopped:
          return "rprof_stopped";
+      case client_events::kRprofCreated:
+         return "rprof_created";
       default:
          LOG_WARNING_MESSAGE("unexpected event type: " + 
                              safe_convert::numberToString(type_));