+++ src/gwt/src/org/rstudio/studio/client/workbench/views/source/editors/text/r/SignatureToolTipManager.java
@@ -15,9 +15,11 @@
 package org.rstudio.studio.client.workbench.views.source.editors.text.r;
 
 import org.rstudio.core.client.Debug;
-import org.rstudio.core.client.DocumentIdleBackgroundTask;
 import org.rstudio.core.client.HandlerRegistrations;
+import org.rstudio.core.client.Rectangle;
 import org.rstudio.core.client.StringUtil;
+import org.rstudio.core.client.dom.DomUtils;
+import org.rstudio.core.client.events.MouseDragHandler.MouseCoordinates;
 import org.rstudio.studio.client.RStudioGinjector;
 import org.rstudio.studio.client.common.codetools.CodeToolsServerOperations;
 import org.rstudio.studio.client.server.ServerError;
@@ -25,13 +27,30 @@ import org.rstudio.studio.client.server.ServerRequestCallback;
 import org.rstudio.studio.client.workbench.prefs.model.UIPrefs;
 import org.rstudio.studio.client.workbench.views.source.editors.text.AceEditor;
 import org.rstudio.studio.client.workbench.views.source.editors.text.DocDisplay;
+import org.rstudio.studio.client.workbench.views.source.editors.text.DocDisplay.AnchoredSelection;
+import org.rstudio.studio.client.workbench.views.source.editors.text.ace.AceEditorNative;
 import org.rstudio.studio.client.workbench.views.source.editors.text.ace.Position;
-import org.rstudio.studio.client.workbench.views.source.editors.text.ace.Range;
+import org.rstudio.studio.client.workbench.views.source.editors.text.ace.Token;
 import org.rstudio.studio.client.workbench.views.source.editors.text.ace.TokenCursor;
 import org.rstudio.studio.client.workbench.views.source.editors.text.events.CursorChangedEvent;
 import org.rstudio.studio.client.workbench.views.source.editors.text.events.CursorChangedHandler;
 
+import com.google.gwt.core.client.Scheduler;
+import com.google.gwt.dom.client.Element;
 import com.google.gwt.dom.client.NativeEvent;
+import com.google.gwt.event.dom.client.BlurEvent;
+import com.google.gwt.event.dom.client.BlurHandler;
+import com.google.gwt.event.dom.client.FocusEvent;
+import com.google.gwt.event.dom.client.FocusHandler;
+import com.google.gwt.event.dom.client.KeyCodes;
+import com.google.gwt.event.logical.shared.AttachEvent;
+import com.google.gwt.event.shared.HandlerRegistration;
+import com.google.gwt.user.client.Event;
+import com.google.gwt.user.client.Event.NativePreviewEvent;
+import com.google.gwt.user.client.Event.NativePreviewHandler;
+import com.google.gwt.user.client.Timer;
+import com.google.gwt.user.client.Window;
+import com.google.gwt.user.client.ui.PopupPanel.PositionCallback;
 import com.google.inject.Inject;
 
 public class SignatureToolTipManager
@@ -43,32 +62,94 @@ public class SignatureToolTipManager
       docDisplay_ = docDisplay;
       handlers_ = new HandlerRegistrations();
       
-      idleTask_ = new DocumentIdleBackgroundTask(
-            docDisplay,
-            new DocumentIdleBackgroundTask.BackgroundTask()
+      timer_ = new Timer()
+      {
+         @Override
+         public void run()
+         {
+            // Bail if we don't ever show tooltips
+            if (!uiPrefs_.showSignatureTooltips().getGlobalValue())
+               return;
+            
+            // Bail if this is a cursor-activated timer and we
+            // don't want idle tooltips
+            if (coordinates_ == null && !uiPrefs_.showFunctionTooltipOnIdle().getGlobalValue())
+               return;
+            
+            resolveActiveFunctionAndDisplayToolTip();
+         }
+      };
+      
+      monitor_ = new Timer()
+      {
+         @Override
+         public void run()
+         {
+            if (ready_)
             {
-               @Override
-               public boolean continueExecution(Position position, boolean isMouse)
-               {
-                  if (!uiPrefs_.showFunctionTooltipOnIdle().getGlobalValue())
-                     return false;
-                  
-                  if (!uiPrefs_.showSignatureTooltips().getGlobalValue())
-                     return false;
-                  
-                  isMouseTriggeredEvent_ = isMouse;
-                  resolveActiveFunctionAndDisplayToolTip(position, !isMouse);
-                  return true;
-               }
-            });
+               timer_.schedule(TIMER_DELAY_MS);
+               ready_ = false;
+            }
+         }
+      };
+      
+      handlers_.add(docDisplay_.addFocusHandler(new FocusHandler()
+      {
+         @Override
+         public void onFocus(FocusEvent event)
+         {
+            attachPreviewHandler();
+            beginMonitoring();
+         }
+      }));
+      
+      handlers_.add(docDisplay_.addBlurHandler(new BlurHandler()
+      {
+         @Override
+         public void onBlur(BlurEvent event)
+         {
+            timer_.cancel();
+            toolTip_.hide();
+            detachPreviewHandler();
+            endMonitoring();
+         }
+      }));
       
       handlers_.add(docDisplay_.addCursorChangedHandler(new CursorChangedHandler()
       {
          @Override
-         public void onCursorChanged(CursorChangedEvent event)
+         public void onCursorChanged(final CursorChangedEvent event)
          {
-            if (isMouseTriggeredEvent_)
-               toolTip_.hide();
+            // Defer so that anchors can update
+            Scheduler.get().scheduleDeferred(new Scheduler.ScheduledCommand()
+            {
+               @Override
+               public void execute()
+               {
+                  // Check to see if the cursor has moved outside of the anchored region.
+                  if (anchor_ != null &&
+                      toolTip_.isShowing() &&
+                      !anchor_.getRange().contains(event.getPosition()))
+                  {
+                     anchor_ = null;
+                     toolTip_.hide();
+                  }
+               }
+            });
+         }
+      }));
+      
+      handlers_.add(toolTip_.addAttachHandler(new AttachEvent.Handler()
+      {
+         @Override
+         public void onAttachOrDetach(AttachEvent event)
+         {
+            if (!event.isAttached())
+            {
+               coordinates_ = null;
+               position_ = null;
+               anchor_ = null;
+            }
          }
       }));
    }
@@ -81,9 +162,63 @@ public class SignatureToolTipManager
       server_ = server;
    }
    
+   private void attachPreviewHandler()
+   {
+      detachPreviewHandler();
+      preview_ = Event.addNativePreviewHandler(new NativePreviewHandler()
+      {
+         @Override
+         public void onPreviewNativeEvent(NativePreviewEvent preview)
+         {
+            if (preview.getTypeInt() == Event.ONMOUSEMOVE)
+            {
+               NativeEvent event = preview.getNativeEvent();
+               coordinates_ = new MouseCoordinates(
+                     event.getClientX(),
+                     event.getClientY());
+               ready_ = true;
+            }
+            else if (preview.getTypeInt() == Event.ONKEYDOWN)
+            {
+               coordinates_ = null;
+               ready_ = true;
+               if (preview.getNativeEvent().getKeyCode() == KeyCodes.KEY_ESCAPE)
+                  suppressed_ = true;
+            }
+         }
+      });
+   }
+   
+   private void detachPreviewHandler()
+   {
+      if (preview_ != null)
+      {
+         preview_.removeHandler();
+         preview_ = null;
+      }
+   }
+   
+   private void beginMonitoring()
+   {
+      if (monitoring_)
+         return;
+      
+      monitor_.scheduleRepeating(MONITOR_DELAY_MS);
+      monitoring_ = true;
+   }
+   
+   private void endMonitoring()
+   {
+      monitor_.cancel();
+      monitoring_ = false;
+   }
+   
    public void detach()
    {
-      idleTask_.detach();
+      detachPreviewHandler();
+      timer_.cancel();
+      endMonitoring();
+      handlers_.removeHandler();
    }
    
    public RCompletionToolTip getToolTip()
@@ -125,21 +260,80 @@ public class SignatureToolTipManager
       });
    }
    
-   public void resolveActiveFunctionAndDisplayToolTip(
-         final boolean searchForFunction)
+   boolean isMouseDrivenEvent()
    {
-      resolveActiveFunctionAndDisplayToolTip(
-            docDisplay_.getCursorPosition(),
-            searchForFunction);
+      return coordinates_ != null;
    }
    
-   public void resolveActiveFunctionAndDisplayToolTip(
-         final Position position,
-         final boolean searchForFunction)
+   Position getLookupPosition()
    {
-            
-      if (!uiPrefs_.showSignatureTooltips().getValue())
+      if (coordinates_ == null)
+      {
+         Position position = docDisplay_.getCursorPosition();
+         
+         // Nudge the cursor column if the cursor currently lies
+         // upon a closing paren.
+         if (docDisplay_.getCharacterAtCursor() == ')' &&
+             position.getColumn() != 0)
+         {
+            position = Position.create(
+                  position.getRow(),
+                  position.getColumn() - 1);
+         }
+         
+         return position;
+      }
+      
+      return docDisplay_.screenCoordinatesToDocumentPosition(
+            coordinates_.getMouseX(),
+            coordinates_.getMouseY());
+   }
+   
+   // Sets an anchored range for a cursor currently lying
+   // on an identifier before a '(' (a function call).
+   private void setAnchor(TokenCursor cursor)
+   {
+      if (anchor_ != null)
+      {
+         anchor_.detach();
+         anchor_ = null;
+      }
+      
+      TokenCursor endCursor = cursor.cloneCursor();
+      if (!endCursor.moveToNextToken())
          return;
+      
+      if (!endCursor.valueEquals("("))
+         return;
+      
+      // TODO: How to anchor if there is no matching ')' for
+      // this function?
+      if (!endCursor.fwdToMatchingToken())
+         return;
+      
+      Position endPos = endCursor.currentPosition();
+      TokenCursor startCursor = cursor.cloneCursor();
+      Token lookbehind = startCursor.peekBwd(1);
+      if (lookbehind.valueEquals("::") || lookbehind.valueEquals(":::"))
+      {
+         if (!startCursor.moveToPreviousToken())
+            return;
+         
+         if (!startCursor.moveToPreviousToken())
+            return;
+      }
+      
+      Position startPos = startCursor.currentPosition();
+      anchor_ = docDisplay_.createAnchoredSelection(startPos, endPos);
+   }
+   
+   public void resolveActiveFunctionAndDisplayToolTip()
+   {
+      if (suppressed_)
+      {
+         suppressed_ = false;
+         return;
+      }
       
       if (docDisplay_.isPopupVisible())
          return;
@@ -148,68 +342,116 @@ public class SignatureToolTipManager
       if (editor == null)
          return;
       
+      final Position position = getLookupPosition();
+      final boolean isMouseEvent = isMouseDrivenEvent();
+      
+      // Ensure that the mouse target is actually the active editor
+      if (isMouseEvent)
+      {
+         Element el = DomUtils.elementFromPoint(
+               coordinates_.getMouseX(),
+               coordinates_.getMouseY());
+         
+         AceEditorNative nativeEditor = AceEditorNative.getEditor(el);
+         if (nativeEditor != null && nativeEditor != editor.getWidget().getEditor())
+            return;
+      }
+      
+      // Hide an older tooltip if this was a mouse move (this allows
+      // the popup to hide if the user hovers over a function name,
+      // and then later moves the mouse away)
+      if (isMouseEvent)
+         toolTip_.hide();
+      
       TokenCursor cursor = editor.getSession().getMode().getRCodeModel().getTokenCursor();
-      if (!cursor.moveToPosition(position))
-         return;
-
-      if (searchForFunction && !cursor.moveToActiveFunction())
+      if (!cursor.moveToPosition(position, true))
          return;
       
+      // If this is a cursor-idle event and the user has opted into
+      // cursor-idle tooltips, then do some extra work to resolve
+      // the location of that function. It's okay if this fails
+      // (implies that perhaps the cursor is already on a function name,
+      // rather than within a function). Note that on failure the cursor
+      // position is not changed.
+      if (!isMouseEvent &&
+          uiPrefs_.showFunctionTooltipOnIdle().getGlobalValue() &&
+          !cursor.valueEquals("("))
+      {
+         cursor.findOpeningBracket("(", false);
+      }
+      
+      Token lookahead = cursor.peekFwd(1);
+      if (lookahead.valueEquals("::") || lookahead.valueEquals(":::"))
+         if (!cursor.moveToNextToken())
+            return;
+      
+      if (cursor.valueEquals("::") || cursor.valueEquals(":::"))
+         if (!cursor.moveToNextToken())
+            return;
+      
+      if (cursor.valueEquals("("))
+         if (!cursor.moveToPreviousToken())
+            return;
+      
+      // Cache the cursor position -- this should be the first
+      // token prior to a '(', forming the active call.
+      // If we already have an active tooltip for the current position,
+      // then bail.
+      if (toolTip_.isShowing() &&
+          cursor.currentPosition().isEqualTo(position_))
+      {
+         return;
+      }
+      position_ = cursor.currentPosition();
+      
+      // Double check that we're in the correct spot for a function call.
+      // The cursor should lie upon an identifier, and the next token should
+      // be an opening paren.
+      if (!cursor.hasType("identifier"))
+         return;
+
       if (!cursor.nextValue().equals("("))
          return;
       
-      final TokenCursor callEndCursor = cursor.cloneCursor();
-      Position callEndPos = callEndCursor.currentPosition();
-      callEndPos.setColumn(
-            callEndPos.getColumn() +
-            callEndCursor.currentValue().length());
-      
-      if (!cursor.findStartOfEvaluationContext())
-         return;
-
-      final String callString = editor.getTextForRange(Range.fromPoints(
-            cursor.currentPosition(),
-            callEndPos));
-      
+      String callString = cursor.currentValue();
       if (isBoringFunction(callString))
          return;
       
-      if (position.equals(lastCursorPosition_))
-         return;
-
-      server_.getArgs(callString, "", new ServerRequestCallback<String>() {
+      // If this is a namespaced function call, then append that context.
+      Token lookbehind = cursor.peekBwd(1);
+      if (lookbehind.valueEquals("::") || lookbehind.valueEquals(":::"))
+      {
+         // Do-while loop just to allow 'break' for control flow
+         do
+         {
+            TokenCursor clone = cursor.cloneCursor();
+            if (!clone.moveToPreviousToken())
+               break;
+            if (!clone.moveToPreviousToken())
+               break;
+            if (!clone.hasType("identifier"))
+               break;
+            callString = clone.currentValue() + "::" + callString;
+         } while (false);
+            
+      }
+      
+      // Set anchor (so we can dismiss popup when cursor moves outside
+      // of anchored region)
+      setAnchor(cursor.cloneCursor());
+      
+      final String fnString = callString;
+      server_.getArgs(fnString, "", new ServerRequestCallback<String>() {
          
          @Override
          public void onResponseReceived(String arguments)
          {
-            final String signature = callString + arguments;
+            final String signature = fnString + arguments;
             
             if (StringUtil.isNullOrEmpty(arguments))
                return;
             
-            if (!searchForFunction)
-            {
-               lastCursorPosition_ = position;
-               toolTip_.resolvePositionAndShow(
-                     signature,
-                     position);
-            }
-            else
-            {
-               TokenCursor endCursor = callEndCursor.cloneCursor();
-               if (!endCursor.moveToNextToken())
-                  return;
-               
-               if (!endCursor.fwdToMatchingToken())
-                  return;
-               
-               lastCursorPosition_ = position;
-               toolTip_.resolvePositionAndShow(
-                     signature,
-                     Range.fromPoints(
-                           callEndCursor.currentPosition(),
-                           endCursor.currentPosition()));
-            }
+            resolvePositionAndShow(signature, position);
          }
 
          @Override
@@ -220,17 +462,49 @@ public class SignatureToolTipManager
       });
    }
    
+   private void resolvePositionAndShow(String signature, Position position)
+   {
+      // Default to displaying before cursor; however, display above
+      // if doing so would place the tooltip too close (or off) of
+      // the window.
+      final Rectangle bounds = docDisplay_.getPositionBounds(position);
+      toolTip_.setText(signature);
+      toolTip_.setPopupPositionAndShow(new PositionCallback()
+      {
+         @Override
+         public void setPosition(int offsetWidth, int offsetHeight)
+         {
+            int left = bounds.getLeft();
+            
+            boolean verticalOverflow = bounds.getBottom() + offsetHeight >= Window.getClientHeight() - 20;
+            int top = verticalOverflow
+                  ? bounds.getTop() - offsetHeight - 10
+                  : bounds.getBottom() + 10;
+                  
+            toolTip_.setPopupPosition(left, top);
+         }
+      });
+   }
+   
    private final RCompletionToolTip toolTip_;
    private final DocDisplay docDisplay_;
    private final HandlerRegistrations handlers_;
-   private final DocumentIdleBackgroundTask idleTask_;
+   private final Timer timer_;
    
-   private boolean isMouseTriggeredEvent_;
-   private Position lastCursorPosition_;
+   private final Timer monitor_;
+   private boolean monitoring_;
+   private boolean suppressed_;
+   
+   private HandlerRegistration preview_;
+   private MouseCoordinates coordinates_;
+   private Position position_;
+   private AnchoredSelection anchor_;
+   private boolean ready_;
 
    private UIPrefs uiPrefs_;
    private CodeToolsServerOperations server_;
    
-   private static final int DELAY_MS = 1200;
+   private static final int MONITOR_DELAY_MS = 200;
+   private static final int TIMER_DELAY_MS   = 900;
 }